Agentic-Travel-Planner/requirements.txt
openai
anthropic
google-generativeai
python-dotenv
httpx
pydantic
fastapi
uvicorn
python-multipart
pypdf
python-docx
langfuse
